requirements.txt
streamlit==1.32.0
pandas==2.2.1
openpyxl==3.1.2
SQLAlchemy==2.0.27
python-dotenv==1.0.1
plotly==5.19.0
numpy==1.26.4
unidecode==1.3.7
psycopg2-binary==2.9.9  # Para PostgreSQL